Feature to create progress bar based on the financial type.
Add financial type to the progress bar interface. Now it is possible to create progress bar
which pull data from the contribution table based on the financial types available in the system.
The contribution pages as well as financial types work together.

Need to add validation on the progress bar creation interface to make the user either select a
financial type or a contribution page.